requirements.txt
disco-py
pymongo
prometheus_client